First naïve version of the naive reallocation
As well as it's vectorial counterpart. The performance ration between
the scalar and vectorial versions of `find` is around 3. Which is great,
but when we use `memcpy` in our reallocation instead of a for loop over
an array of int we actually are 70 times faster (and we assume this
`memcpy` isn't even vectorial).