Put utility functions in separate .h and .c files
The utility functions such as `print_array` or `generate_array` have
been put in separate files. To make compilation easier, a Makefile has
been added to the project. The build now goes under `gcc_build` (GCC is
precised here because we thought it could be interesting to compile with
CLang and ICC and see wether or not the results we are getting differ).

The functions related to argument parsing have also been put in separate
files to clean up the main file.

Finally a `.ycm_extra_conf.py` file has been added to enable us to use
YouCompleteMe: a cutting edge C autocompletion tool that introspecs the
code (by actually compiling it... with Clang) and suggests functions
from the imported header files, displays their documentation if they
have some, enable us to `GoTo` their definition... Every Vim user should
be using this plugin even though it's a pain to setup :)