Last round of optimisization for vect_find
The vectorial implementation of `find` has been slightly boosted by
inlining the vectorial comparisons instead of putting them in a `for`
loop. For introduces 9 comparisons and 9 incrementations while with
inlining - since we know we have 8 elements to process per vector - we
can get rid of the comparisons and significantly improve the
performance.

We also got rid of one variable assigment which seemed to be costly
(maybe because that variable had to be fetched from the L3 cache...).

EDIT:
Having a look at the assembly code produced we can indeed confirm that
one instruction has been removed.